# dataset loading speed: Inline imports