Moves runtime imports to top of file and narrow down.
Should help (a bit) everywhere these functions are used.

After
```
time PYTHONPATH=. STEPS=2 WINO=0 kernprof -lv examples/hlb_cifar10.py
  0 1425.81 ms run, 1425.59 ms python,    0.23 ms CL, 1199.34 loss, 0.002262 LR, 0.84 GB used,    473.17 GFLOPS
  1  652.53 ms run,  652.34 ms python,    0.19 ms CL, 1196.44 loss, 0.000086 LR, 3.17 GB used,   1033.90 GFLOPS
Wrote profile results to hlb_cifar10.py.lprof
Timer unit: 1e-06 s


real    0m11.992s
user    0m10.608s
sys     0m2.488s
```

Before
```
time PYTHONPATH=. STEPS=2 WINO=0 kernprof -lv examples/hlb_cifar10.py
  0 1395.06 ms run, 1394.83 ms python,    0.23 ms CL, 1199.34 loss, 0.002262 LR, 0.84 GB used,    483.60 GFLOPS
  1  649.72 ms run,  649.49 ms python,    0.23 ms CL, 1196.44 loss, 0.000086 LR, 3.17 GB used,   1038.37 GFLOPS
Wrote profile results to hlb_cifar10.py.lprof
Timer unit: 1e-06 s


real    0m12.631s
user    0m11.092s
sys     0m2.742s
```