@geohot 2023-10-07
In fact, I think all the fake_torch_load stuff can be deleted! convnext looks like the only thing that hasn't moved